backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
duckdb==1.1.3
pyarrow==18.1.0
pandas==2.2.3
google-cloud-storage==2.19.0
groq==0.13.0
nba_api==1.5.2
python-dotenv==1.0.1
httpx==0.27.2
